Fixed various problems with the Scratch import, including:
- Costumes and sounds effectively have a different namespace for each sprite.  So different sprites can (and usually do) have costumes named "costume1" that refer to different images.  Now the assets are prefixed with the name of the class they belong to, to avoid clashes.
- Image sizes were sometimes being set using floating point numbers rather than integers, which caused a compile-time error in the generated code.
- Scratch sprites can be called things like "World", which causes some confusion (and errors!).  The current solution is to special-case World and Actor, and rename any such sprites
- Generally improved some of the generated code to avoid problems that were causing compile-time errors in the imported scenario.

(from subversion r8508)

+++ greenfoot/src/greenfoot/importer/scratch/ScratchSpriteMorph.java
@@ -49,9 +49,9 @@ public class ScratchSpriteMorph extends ScriptableScratchMorph
     {
         acc.append("GreenfootImage img = getImage();\n");
         acc.append("img.scale(")
-           .append(getBounds().getWidth())
+           .append(getBounds().getWidth().intValue())
            .append(", ")
-           .append(getBounds().getHeight())
+           .append(getBounds().getHeight().intValue())
            .append(");\n");
     }
